Fixed Immortality action, Updated Shield replenish action.
Fixed Immortality Action, added leveling logic for Shield replenish.

BUG: removing an action module will not reset ActionData. This will case issues with actions that need to make changes to ActionData as you will be getting edited values upon adding the action back on.

+++ Assets/Scripts/Scriptables/Systems/ImmortalitySystem.cs
@@ -15,10 +15,13 @@ namespace Celeritas.Scriptables.Systems
 
 	public class ImmortalitySystem : ModifierSystem
 	{
+		/// <inheritdoc/>
 		public override bool Stacks => true;
 
+		/// <inheritdoc/>
 		public override SystemTargets Targets => SystemTargets.Ship;
 
-		public override string GetTooltip(ushort level) => $"Grants temporary immortality.";
+		/// <inheritdoc/>
+		public override string GetTooltip(int level) => $"Grants temporary immortality.";
 	}
 }
